CLI can now parse link
